chore: extract desktop Filter to a new component
