+++ packages/core/src/components/search/Filter/Filter.tsx
@@ -1,16 +1,8 @@
 import { gql } from '@faststore/graphql-utils'
-import { setFacet, toggleFacet, useSearch } from '@faststore/sdk'
 
-import {
-  FacetBoolean as UIFacetBoolean,
-  FacetBooleanItem as UIFacetBooleanItem,
-  FacetRange as UIFacetRange,
-  Facets as UIFacets,
-  Filter as UIFilter,
-  useUI,
-} from '@faststore/ui'
+import { useUI } from '@faststore/ui'
 import type { Filter_FacetsFragment } from '@generated/graphql'
-import { useFormattedPrice } from 'src/sdk/product/useFormattedPrice'
+import FilterFixed from './FilterFixed'
 import FilterSlider from './FilterSlider'
 import { useFilter } from './useFilter'
 
@@ -28,83 +20,12 @@ interface Props {
 
 function Filter({ facets: allFacets, testId = 'fs-filter' }: Props) {
   const filter = useFilter(allFacets)
-  const { resetInfiniteScroll, state, setState } = useSearch()
   const { filter: displayFilter } = useUI()
-  const { facets, expanded, dispatch } = filter
 
   return (
     <>
       <div className="hidden-mobile">
-        <UIFilter
-          testId={`desktop-${testId}`}
-          title="Filters"
-          indicesExpanded={expanded}
-          onAccordionChange={(idx) =>
-            dispatch({ type: 'toggleExpanded', payload: idx })
-          }
-        >
-          {facets.map((facet, index) => {
-            const { __typename: type, label } = facet
-            const isExpanded = expanded.has(index)
-            return (
-              <UIFacets
-                key={`${testId}-${label}-${index}`}
-                testId={testId}
-                index={index}
-                type={type}
-                label={label}
-              >
-                {type === 'StoreFacetBoolean' && isExpanded && (
-                  <UIFacetBoolean>
-                    {facet.values.map((item) => (
-                      <UIFacetBooleanItem
-                        key={`${testId}-${facet.label}-${item.label}`}
-                        id={`${testId}-${facet.label}-${item.label}`}
-                        testId={testId}
-                        onFacetChange={(facet) => {
-                          setState({
-                            ...state,
-                            selectedFacets: toggleFacet(
-                              state.selectedFacets,
-                              facet
-                            ),
-                            page: 0,
-                          })
-                          resetInfiniteScroll(0)
-                        }}
-                        selected={item.selected}
-                        value={item.value}
-                        quantity={item.quantity}
-                        facetKey={facet.key}
-                        label={item.label}
-                      />
-                    ))}
-                  </UIFacetBoolean>
-                )}
-                {type === 'StoreFacetRange' && isExpanded && (
-                  <UIFacetRange
-                    facetKey={facet.key}
-                    min={facet.min}
-                    max={facet.max}
-                    formatter={useFormattedPrice}
-                    onFacetChange={(facet) => {
-                      setState({
-                        ...state,
-                        selectedFacets: setFacet(
-                          state.selectedFacets,
-                          facet,
-                          true
-                        ),
-                        page: 0,
-                      })
-                      resetInfiniteScroll(0)
-                    }}
-                  />
-                )}
-              </UIFacets>
-            )
-          })}
-        </UIFilter>
+        <FilterFixed {...filter} testId={testId} title="Filters" />
       </div>
 
       {displayFilter && (
